wip: externalize CoinjoinPreconditionViolationAlert component
